chore: Update npm dependencies and add new packages
This commit updates the npm dependencies for the project. It includes updating the `jsonwebtoken` package to version 9.0.2 and adding the `buffer-equal-constant-time` and `ecdsa-sig-formatter` packages. These updates and additions ensure that the project is using the most recent stable versions of the packages, which may include bug fixes or performance improvements.